change mechanism for `deps`; add `depsRequired`; add `depsFrom`.
With this commit, the user is forced to explicitly specify all required dependency names via `depsRequired`.
`options.deps` contains generated options for all names of `depsRequired`.
These options are populated with defaults from `depsFrom`,
The examples under `./examples` show how this is used.

The `coercedTo` type of `deps` had some weaknesses:
1. It was not possible to specify required dependencies explicitly, because it was not possible to create nested options like`options.deps.bash = mkOption ...`.
2. Because of (1.) I had to add a manual check in makeModule.nix to ensure that all required `deps` are populated by the user.

Having actual options for all required dependencies allows to reduce complexity and hand the `checking` off to the module system.

Also, when rendering the options for a package to a doc, all dependencies now appear as an option, which I think is great. It allows the user to see all inputs of a package without having to read the code.