feat: allow options for `deps`
This changes the type for `deps` so that nested options are allowed.

A drv module can now specify required dependencies.
The module system will ensure that those options will be set.
This also allows to render a manual for a packages interface including the required deps.

The `deps = {pkgs, ...}: {inherit (deps) ...;}` style remains intact.